Update render.R
update render to be based on parameters. Anyone using this repo should be able to declare variables for their sheet, link to their github hosted page, and the location of the pdf version on github. This is based on assumption that github pages has been set up with a <username>.github.io repo